Initial Commit Of All Files
This is a PowerShell script that communicates with a DJI Tello drone. Using the .NET Speech.Synthesizer, it listens on the default OS mic and speaks over the default OS speakers. Flight control commands are sent via UDP to a static IP after a key word is detected by the Speech Engine.